Move binding logic into respective utils
